Navbar and card updates
